feat: 찜 버튼 분리
